backend/app/requirements.txt
anyio==3.7.0
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
backcall==0.2.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contextlib2==21.6.0
cryptography==41.0.1
decorator==5.1.1
dill==0.3.6
Django==4.2.2
django-ninja==0.22.2
django-ninja-extra==0.19.1
django-ninja-jwt==5.2.5
django-phonenumber-field==7.1.0
executing==1.2.0
fcm-django==2.0.0
firebase-admin==6.2.0
flake8==6.0.0
flet==0.7.4
flet-core==0.7.4
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.4
injector==0.20.1
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.16
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyparsing==3.1.0
python-environ==0.4.54
python-slugify==8.0.1
redis==4.6.0
repath==0.9.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
tomlkit==0.11.8
traitlets==5.9.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
watchdog==2.3.1
wcwidth==0.2.6
websocket-client==1.6.1
websockets==10.4
wrapt==1.15.0
